Enhance photo gallery functionality and user experience. Introduce empty gallery fallback with setup instructions, implement loading states for gallery tiles, and improve photo preview features including zoom and back navigation. Refactor photo handling into a dedicated composable for better organization and maintainability. Add tests for EXIF data formatting and ensure metadata privacy by omitting GPS coordinates. Update documentation to reflect these changes.
